+++ DashboardUI/ChartRequest.cs
@@ -10,6 +10,6 @@ namespace DashboardUI
     {
         public List<double> Months { get; set; }
         public List<string> Legends { get; set; }
-        public List<List<int>> Series { get; set; }
+        public List<List<double>> Series { get; set; }
     }
 }
