+++ Entities/Concrete/DepartmentCapacity.cs
@@ -13,7 +13,7 @@ namespace Entities.Concrete
         [Key]
         public int DCapacityId { get; set; }
         public int DepartmentId { get; set; }
-        public int DTotalCapacity { get; set; }
+        public double DTotalCapacity { get; set; }
         public DateTime Date { get; set; }
     }
 }
