+++ Entities/Concrete/ProjectCapacity.cs
@@ -11,7 +11,7 @@ namespace Entities.Concrete
     {
         public int ProjectCapacityId { get; set; }
         public int ProjectId { get; set; }
-        public int PTotalCapacity { get; set; }
+        public double PTotalCapacity { get; set; }
         public DateTime Date { get; set; }
     }
 }
